All new workflows and data structures
This commit contains results from work responding to initial reviews. It includes a full new execution of the latest logic in the bispy package, creating some additional standardization in the data structures. There is a new dataset of ScienceBase Data Release products included and a new workflow that outputs a flattened CSV summarization of select properties for other types of uses.